+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add creation of `read-releases` ClusterRole.
+
 ## [0.21.0] - 2022-02-16
 
 ### Added